added tests

+++ matrix-spring-boot-bot/src/main/kotlin/net/folivo/matrix/bot/appservice/membership/AppserviceMembershipChangeService.kt
@@ -3,6 +3,7 @@ package net.folivo.matrix.bot.appservice.membership
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.map
 import net.folivo.matrix.bot.appservice.room.MatrixRoomService
+import net.folivo.matrix.bot.appservice.user.MatrixUser
 import net.folivo.matrix.bot.appservice.user.MatrixUserService
 import net.folivo.matrix.bot.membership.MembershipChangeService
 import net.folivo.matrix.bot.util.BotServiceHelper
@@ -10,7 +11,7 @@ import net.folivo.matrix.restclient.MatrixClient
 import org.slf4j.LoggerFactory
 import org.springframework.transaction.annotation.Transactional
 
-class DefaultAppserviceMembershipChangeService(
+class AppserviceMembershipChangeService(
         private val roomService: MatrixRoomService,
         private val membershipService: MatrixMembershipService,
         private val userService: MatrixUserService,
@@ -23,21 +24,31 @@ class DefaultAppserviceMembershipChangeService(
     }
 
     @Transactional
-    override suspend fun onRoomJoin(roomId: String, userId: String) {
+    override suspend fun onRoomJoin(userId: String, roomId: String) {
         LOG.debug("saveRoomJoin in room $roomId of user $userId")
         roomService.getOrCreateRoom(roomId)
         userService.getOrCreateUser(userId)
-        membershipService.getOrCreateMembership(userId, roomId)
+        membershipService.getOrCreateMembership(userId = userId, roomId = roomId)
     }
 
     @Transactional
-    override suspend fun onRoomLeave(roomId: String, userId: String) {
-        val membershipsSize = membershipService.getMembershipsSizeByRoomId(roomId)
-        if (membershipsSize > 1) {
+    override suspend fun onRoomLeave(userId: String, roomId: String) {
+        // we do this here because it also syncs the room if it is not in database
+        val room = roomService.getOrCreateRoom(roomId)
+
+        if (membershipService.doesRoomContainsMembers(roomId, setOf(userId))) {
+
+            val user = userService.getOrCreateUser(userId)
+
             LOG.debug("save room leave in room $roomId of user $userId")
             membershipService.deleteMembership(userId, roomId)
 
-            if (membershipService.hasRoomOnlyManagedUsersLeft(roomId)) {
+            deleteUserWhenNotManaged(user)
+
+            val noMembersLeft = membershipService.getMembershipsSizeByRoomId(roomId) == 0L
+            val onlyManagedUsersLeft = membershipService.hasRoomOnlyManagedUsersLeft(roomId)
+
+            if (onlyManagedUsersLeft) {
                 LOG.debug("leave room $roomId with all managed users because there are only managed users left")
                 val memberships = membershipService.getMembershipsByRoomId(roomId)
                 memberships
@@ -46,20 +57,25 @@ class DefaultAppserviceMembershipChangeService(
                             if (joinedUserId == helper.getBotUserId())
                                 matrixClient.roomsApi.leaveRoom(roomId)
                             else matrixClient.roomsApi.leaveRoom(roomId, joinedUserId)
+                            membershipService.deleteMembership(joinedUserId, roomId)
+                            deleteUserWhenNotManaged(userService.getOrCreateUser(joinedUserId))
                         }
             }
-        } else {
-            LOG.debug("delete room $roomId and membership because there are no members left")
-            membershipService.deleteMembership(userId, roomId)
-            roomService.deleteRoom(roomId)
-            if (membershipService.getMembershipsSizeByUserId(userId) == 0L) {
-                LOG.debug("delete user $userId because there are no memberships left")
-                userService.deleteUser(userId)
+            if (!room.isManaged && (onlyManagedUsersLeft || noMembersLeft)) {
+                roomService.deleteRoom(roomId)
             }
+
         }
     }
 
-    override suspend fun shouldJoinRoom(roomId: String, userId: String): Boolean {
+    private suspend fun deleteUserWhenNotManaged(user: MatrixUser) {
+        if (!user.isManaged && membershipService.getMembershipsSizeByUserId(user.id) == 0L) {
+            LOG.debug("delete user ${user.id} because there are no memberships left")
+            userService.deleteUser(user.id)
+        }
+    }
+
+    override suspend fun shouldJoinRoom(userId: String, roomId: String): Boolean {
         return true
     }
 }
